+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [CalVer](https://calver.org/).
 - WFS-T filter parser will now drop namespaces in property names. If not this is done, the resulting SQL will be invalid.
 - Bug regarding no-strip of gml namespace on Envelope BBOX filter is fixed.
 - Fixed bug in Feature API regarding namespace changes in WFS-t.
+- Trim double qoutes from ogc:PropertyName in WFS-t. Openlayers adds them in WFS requets. 
+- Always set ns uri to http:// in WFS-t or else editing won't work in Admin.
+
 
 ## [2021.5.1]
 ### Added